requirements.txt
streamlit
pandas
plotly
pyyaml
requests
python-dotenv
